Check distribution before architecture

+++ source/_static/install_dependencies.sh
@@ -80,28 +80,30 @@ if [[ $(uname) == "Darwin" ]]; then
     echo
     echo "=== Installed successfully!  IMPORTANT: For changes to take effect,"
     echo "please close and reopen the terminal window, or run:  exec \$SHELL"
-elif [[ ! $(uname -m) =~ ^arm* ]]; then
+elif [ -f /etc/os-release ]; then
     # shellcheck source=/etc/os-release
     source /etc/os-release
     case "$NAME" in
-    Ubuntu)
-        sudo apt-get update
-        sudo apt-get install -y "${ubuntu_pkgs[@]}"
-        echo 'SUBSYSTEM=="usb", ATTRS{idVendor}=="03e7", MODE="0666"' | sudo tee /etc/udev/rules.d/80-movidius.rules
-        sudo udevadm control --reload-rules && sudo udevadm trigger
-        python3 -m pip install --upgrade pip
+    Ubuntu|Debian*)
+        if [[ ! $(uname -m) =~ ^arm* ]]; then
+            sudo apt-get update
+            sudo apt-get install -y "${ubuntu_pkgs[@]}"
+            echo 'SUBSYSTEM=="usb", ATTRS{idVendor}=="03e7", MODE="0666"' | sudo tee /etc/udev/rules.d/80-movidius.rules
+            sudo udevadm control --reload-rules && sudo udevadm trigger
+            python3 -m pip install --upgrade pip
+        elif [[ $(uname -m) =~ ^arm* ]]; then
+            sudo apt-get update
+            sudo apt-get install -y "${ubuntu_pkgs[@]}" "${ubuntu_arm_pkgs[@]}"
+            echo 'SUBSYSTEM=="usb", ATTRS{idVendor}=="03e7", MODE="0666"' | sudo tee /etc/udev/rules.d/80-movidius.rules
+            sudo udevadm control --reload-rules && sudo udevadm trigger
+            python3 -m pip install --upgrade pip
+        fi
         ;;
     *)
         echo "ERROR: Distribution not supported"
         exit 99
         ;;
     esac
-elif [[ $(uname -m) =~ ^arm* ]]; then
-    sudo apt-get update
-    sudo apt-get install -y "${ubuntu_pkgs[@]}" "${ubuntu_arm_pkgs[@]}"
-    echo 'SUBSYSTEM=="usb", ATTRS{idVendor}=="03e7", MODE="0666"' | sudo tee /etc/udev/rules.d/80-movidius.rules
-    sudo udevadm control --reload-rules && sudo udevadm trigger
-    python3 -m pip install --upgrade pip
 else
     echo "ERROR: Host not supported"
     exit 99
